requirements.txt
flake8==7.2.0
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.1.0
django-phonenumber-field==8.0.0
celery==5.4.0
requests==2.32.3
djangorestframework_simplejwt==5.5.0
django-celery-beat==2.7.0
django-cors-headers==4.7.0
redis==5.2.1
drf-yasg==1.21.10
phonenumbers==9.0.1

